With harness

+++ src/planners/src/wrappers/trajopt/RobotModelWrapper.cpp
@@ -3,14 +3,15 @@
 
 typedef Eigen::Matrix<float, 3, 4> Matrix4f;
 
-
 template <typename TElem>
-std::ostream& operator<<(std::ostream& os, const std::vector<TElem>& vec) {
+std::ostream &operator<<(std::ostream &os, const std::vector<TElem> &vec)
+{
     typedef typename vector<TElem>::const_iterator iter_t;
     const iter_t iter_begin = vec.begin();
-    const iter_t iter_end   = vec.end();
+    const iter_t iter_end = vec.end();
     os << "[";
-    for (iter_t iter = iter_begin; iter != iter_end; ++iter) {
+    for (iter_t iter = iter_begin; iter != iter_end; ++iter)
+    {
         std::cout << ((iter != iter_begin) ? "," : "") << *iter;
     }
     os << "]";
@@ -18,16 +19,17 @@ std::ostream& operator<<(std::ostream& os, const std::vector<TElem>& vec) {
 }
 
 template <typename Map>
-bool key_compare (Map const &lhs, Map const &rhs) {
+bool key_compare(Map const &lhs, Map const &rhs)
+{
 
-    auto pred = [] (decltype(*lhs.begin()) a, decltype(a) b)
+    auto pred = [](decltype(*lhs.begin()) a, decltype(a) b)
     { return a.first == b.first; };
 
-    return lhs.size() == rhs.size()
-            && std::equal(lhs.begin(), lhs.end(), rhs.begin(), pred);
+    return lhs.size() == rhs.size() && std::equal(lhs.begin(), lhs.end(), rhs.begin(), pred);
 }
 
-void skew(Eigen::Vector3d& v, Eigen::Matrix3d* result) {
+void skew(Eigen::Vector3d &v, Eigen::Matrix3d *result)
+{
     /*
     Skew-symmetric matrix:
     A^T = −A.
@@ -48,37 +50,44 @@ void skew(Eigen::Vector3d& v, Eigen::Matrix3d* result) {
     (*result)(2, 2) = 0.0;
 }
 
-void mul(const Eigen::Matrix3d &a, const Eigen::Matrix3Xd &b, Eigen::Matrix3Xd *result) {
-    if (b.cols() != result->cols()) {
-        std::cerr<<"size missmatch. b.cols()= " << static_cast<int>(b.cols()) << "result->cols()= " << static_cast<int>(result->cols()) << std::endl;
+void mul(const Eigen::Matrix3d &a, const Eigen::Matrix3Xd &b, Eigen::Matrix3Xd *result)
+{
+    if (b.cols() != result->cols())
+    {
+        std::cerr << "size missmatch. b.cols()= " << static_cast<int>(b.cols()) << "result->cols()= " << static_cast<int>(result->cols()) << std::endl;
     }
 
-    for (int col = 0; col < b.cols(); col++) {
-        const double x = a(0,0)*b(0,col)+a(0,1)*b(1,col)+a(0,2)*b(2,col);
-        const double y = a(1,0)*b(0,col)+a(1,1)*b(1,col)+a(1,2)*b(2,col);
-        const double z = a(2,0)*b(0,col)+a(2,1)*b(1,col)+a(2,2)*b(2,col);
+    for (int col = 0; col < b.cols(); col++)
+    {
+        const double x = a(0, 0) * b(0, col) + a(0, 1) * b(1, col) + a(0, 2) * b(2, col);
+        const double y = a(1, 0) * b(0, col) + a(1, 1) * b(1, col) + a(1, 2) * b(2, col);
+        const double z = a(2, 0) * b(0, col) + a(2, 1) * b(1, col) + a(2, 2) * b(2, col);
         (*result)(0, col) = x;
         (*result)(1, col) = y;
         (*result)(2, col) = z;
     }
 }
 
-void sub(const Eigen::Matrix3Xd &a, const Eigen::Matrix3Xd &b, Eigen::Matrix3Xd *result) {
-    if (a.cols() != b.cols()) {
-        std::cerr<<"size missmatch. a.cols()= "<< static_cast<int>(a.cols()) <<"b.cols()= " << static_cast<int>(b.cols());
+void sub(const Eigen::Matrix3Xd &a, const Eigen::Matrix3Xd &b, Eigen::Matrix3Xd *result)
+{
+    if (a.cols() != b.cols())
+    {
+        std::cerr << "size missmatch. a.cols()= " << static_cast<int>(a.cols()) << "b.cols()= " << static_cast<int>(b.cols());
     }
-    for (int col = 0; col < b.cols(); col++) {
-        for (int row = 0; row < 3; row++) {
+    for (int col = 0; col < b.cols(); col++)
+    {
+        for (int row = 0; row < 3; row++)
+        {
             (*result)(row, col) = a(row, col) - b(row, col);
         }
     }
 }
 
-void RobotModelWrapper::dblVecToEigenVectorXd(const DblVec &in, Eigen::VectorXd &out){
+void RobotModelWrapper::dblVecToEigenVectorXd(const DblVec &in, Eigen::VectorXd &out)
+{
     out = Eigen::VectorXd::Map(in.data(), in.size());
 }
-RobotModelWrapper::RobotModelWrapper(std::shared_ptr<RobotModel> &robot_model):
-    m_robot_model(robot_model)
+RobotModelWrapper::RobotModelWrapper(std::shared_ptr<RobotModel> &robot_model) : m_robot_model(robot_model)
 {
     m_planning_group_name_ = robot_model->getPlanningGroupName();
     robot_model->getPlanningGroupJointsName(robot_model->getPlanningGroupName(), m_planning_group_joints_names_);
@@ -91,17 +100,19 @@ void RobotModelWrapper::setRobotModel(std::shared_ptr<RobotModel> robot_model)
     robot_model->getPlanningGroupJointsName(robot_model->getPlanningGroupName(), m_planning_group_joints_names_);
 }
 
-void RobotModelWrapper::setDOFValues(const DblVec &dofs) {
+void RobotModelWrapper::setDOFValues(const DblVec &dofs)
+{
     Eigen::VectorXd joint_values;
     dblVecToEigenVectorXd(dofs, joint_values);
     m_robot_model->updateJointGroup(m_planning_group_joints_names_, joint_values);
-
 }
 
-void RobotModelWrapper::getDOFLimits(DblVec &lower, DblVec &upper) const {
+void RobotModelWrapper::getDOFLimits(DblVec &lower, DblVec &upper) const
+{
     m_robot_model->getJointLimits(lower, upper);
 }
-void RobotModelWrapper::getDOFValues(std::map<std::string, double> &joint_states_map) {
+void RobotModelWrapper::getDOFValues(std::map<std::string, double> &joint_states_map)
+{
     std::vector<std::string> joint_names(m_planning_group_joints_names_.size());
     DblVec joint_values(m_planning_group_joints_names_.size());
     getDOFValues(joint_names, joint_values);
@@ -110,17 +121,20 @@ void RobotModelWrapper::getDOFValues(std::map<std::string, double> &joint_states
         joint_states_map[joint_names.at(i)] = joint_values.at(i);
 }
 
-void RobotModelWrapper::getDOFValues(vector<std::string> &joint_names, DblVec &joint_values) {
+void RobotModelWrapper::getDOFValues(vector<std::string> &joint_names, DblVec &joint_values)
+{
     joint_names.resize(m_planning_group_joints_names_.size());
     joint_values.resize(m_planning_group_joints_names_.size());
 
-    for (int i = 0; i < m_planning_group_joints_names_.size(); i++){
+    for (size_t i = 0; i < m_planning_group_joints_names_.size(); i++)
+    {
         joint_names.at(i) = m_planning_group_joints_names_.at(i);
         joint_values.at(i) = m_robot_model->getRobotState().robot_joints_[m_planning_group_joints_names_.at(i)].getJointValue();
     }
 }
 
-DblVec RobotModelWrapper::getDOFValues() {
+DblVec RobotModelWrapper::getDOFValues()
+{
     vector<std::string> joint_names(m_planning_group_joints_names_.size());
     DblVec joint_values(m_planning_group_joints_names_.size());
 
@@ -128,32 +142,33 @@ DblVec RobotModelWrapper::getDOFValues() {
     return joint_values;
 }
 
-int RobotModelWrapper::getDOF() const {
+int RobotModelWrapper::getDOF() const
+{
     return m_planning_group_joints_names_.size();
 }
 
-DblMatrix RobotModelWrapper::getPositionJacobian(std::string link_name, const Vector3d &pt) /*const*/ {
-
+DblMatrix RobotModelWrapper::getPositionJacobian(std::string link_name, const Vector3d &pt) /*const*/
+{
 
     DblMatrix m /*=  Eigen::MatrixXd::Zero()*/;
     m.resize(3, getDOF());
     m *= 0.000;
 
-    std::vector< std::pair<std::string,urdf::Joint> > planning_groups_joints;
+    std::vector<std::pair<std::string, urdf::Joint>> planning_groups_joints;
     std::string base_link = m_robot_model->getBaseFrameName();
     std::string tip_link = m_robot_model->getTipFrameName();
 
     m_robot_model->getPlanningGroupJointInformation(m_planning_group_name_, planning_groups_joints);
 
-    KDL::Jacobian  jacobian;
+    KDL::Jacobian jacobian;
 
     std::map<std::string, double> joint_states_map;
 
     m_robot_model->getChainJointState(base_link, link_name, joint_states_map);
     m_robot_model->computeJacobain(base_link, link_name, joint_states_map, jacobian);
-//    jacobian.changeRefPoint(KDL::Vector(pt.x(), pt.y(), pt.z()));
+    //    jacobian.changeRefPoint(KDL::Vector(pt.x(), pt.y(), pt.z()));
 
-    m.block(0,0,3, joint_states_map.size()) = jacobian.data.block(0,0,3,  joint_states_map.size());
+    m.block(0, 0, 3, joint_states_map.size()) = jacobian.data.block(0, 0, 3, joint_states_map.size());
 
     KDL::Frame link_frame = m_robot_model->getRobotState().robot_links_[link_name].getLinkFrame();
 
@@ -172,7 +187,7 @@ DblMatrix RobotModelWrapper::getPositionJacobian(std::string link_name, const Ve
 
     jac_rot *= 0.00;
 
-    jac_rot.block(0,0,3, joint_states_map.size()) = jacobian.data.block(3,0,3,  joint_states_map.size());
+    jac_rot.block(0, 0, 3, joint_states_map.size()) = jacobian.data.block(3, 0, 3, joint_states_map.size());
 
     Eigen::Matrix3Xd jac_rot1 /*=  Eigen::MatrixXd::Zero()*/;
     jac_rot1.resize(3, getDOF());
@@ -192,25 +207,27 @@ DblMatrix RobotModelWrapper::getPositionJacobian(std::string link_name, const Ve
     jac_new1 = jac_new;
 
     return jac_new1;
-
 }
 
-//DblMatrix RobotModelWrapper::getRotationJacobian(std::string link_name) const {}
+// DblMatrix RobotModelWrapper::getRotationJacobian(std::string link_name) const {}
 
-DblVec RobotModelWrapper::setRandomDOFValues() {
+DblVec RobotModelWrapper::setRandomDOFValues()
+{
     std::map<std::string, double> planning_groups_joints_with_random_values;
 
     m_robot_model->generateRandomJointValue(m_planning_group_name_, planning_groups_joints_with_random_values);
 
     DblVec random_joint_values(m_planning_group_joints_names_.size());
 
-    for (int i; i < m_planning_group_joints_names_.size(); i++){
+    for (size_t i; i < m_planning_group_joints_names_.size(); i++)
+    {
         random_joint_values.at(i) = planning_groups_joints_with_random_values[m_planning_group_joints_names_[i]];
     }
     return random_joint_values;
 }
 
-geometry::Transform RobotModelWrapper::getLinkTransformByName(std::string link_name){
+geometry::Transform RobotModelWrapper::getLinkTransformByName(std::string link_name)
+{
 
     Eigen::Vector3d pos;
     Eigen::Vector4d orn;
@@ -218,13 +235,13 @@ geometry::Transform RobotModelWrapper::getLinkTransformByName(std::string link_n
     m_robot_model->getLinkTransformByName(link_name, pos, orn);
 
     return geometry::Transform(geometry::Vector(orn[0], orn[1], orn[2], orn[3]), geometry::Vector(pos[0], pos[1], pos[2]));
-
 }
 
 bool RobotModelWrapper::checkIfLinkExists(const std::string link_name)
 {
-//    if ( m_robot_model->getRobotState().robot_links_.find(link_name) != m_robot_model->getRobotState().robot_links_.end() ) {
-    if (m_robot_model->getRobotState().robot_links_.count(link_name)){
+    //    if ( m_robot_model->getRobotState().robot_links_.find(link_name) != m_robot_model->getRobotState().robot_links_.end() ) {
+    if (m_robot_model->getRobotState().robot_links_.count(link_name))
+    {
         return true;
     }
